+++ CHANGELOG.md
@@ -7,6 +7,7 @@
   - Bump clj-kondo to `2023.06.02-20230612.113558-5`.
   - Improve support for ClojureDart `.cljd` files. #1589
   - Slightly reduce OutOfMemory exceptions that may happen with the JVM version.
+  - Support ignore unused-public-vars via `:linters :clojure-lsp/unused-public-var :exclude-when-contains-meta` clj-kondo setting.
   
 - Editor
   - New code actions: `Replace ':refer :all' with ':refer [my-refer]'` and `Replace ':refer :all' with alias`. #1575